framework_python/test_py_web/requirements.txt
selenium
webdriver-manager
jsonschema
pytest
pytest-bdd
pydash
openpyxl
pytest-xdist
allure-pytest
allure-python-commons
retrying
xlsxwriter
pytest-rerunfailures
pyodbc
pandas